Move NoContentError in pages folder
